pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "distillery"
version = "0.1.0"
description = "A configurable embedding system with DuckDB storage and MCP server integration"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Distillery Contributors"}
]
license = {text = "Apache-2.0"}
keywords = ["knowledge-base", "second-brain", "embeddings", "mcp", "claude-code", "vector-search", "duckdb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "duckdb>=0.8.0",
    "pyyaml>=6.0",
    "httpx>=0.24.0",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
    "pytest-cov>=4.0",
    "types-PyYAML>=6.0",
]
eval = [
    "anthropic>=0.40.0",
]

[project.scripts]
distillery = "distillery.cli:main"
distillery-mcp = "distillery.mcp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/distillery"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = ["-v", "--strict-markers", "--tb=short"]
markers = [
    "unit: unit tests",
    "integration: integration tests",
    "eval: Claude-powered skill evaluation tests (requires ANTHROPIC_API_KEY)",
]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["distillery"]

[tool.coverage.run]
source = ["src/distillery"]
omit = [
    "src/distillery/mcp/__main__.py",
    "src/distillery/store/__main__.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
